requirements.txt
fastapi
uvicorn
passlib
bcrypt
sqlalchemy
python-multipart
aiofiles
python-jose